+++ .gitignore
@@ -1,1 +1,3 @@
 output_file
+target_file_name
+data
