+++ internal/utils/guid.go
@@ -1,0 +1,7 @@
+package utils
+
+import "github.com/google/uuid"
+
+func GenerateSessionID() string {
+	return uuid.New().String()
+}
