--- internal/utils/utils.go
@@ -1,7 +1,0 @@
-package utils
-
-import "github.com/google/uuid"
-
-func GenerateSessionID() string {
-	return uuid.New().String()
-}
